refactor(core): extract hp_bonus and expertise from progression
Break feats↔progression circular imports and slim the progression module.
