+++ core/races.py
@@ -162,7 +162,7 @@ def get_racial_hp_bonus_sources(
     race_id: str, subrace_id: str | None = None
 ) -> list[Any]:
     """Именованные бонусы HP за уровень из grants расы/подрасы."""
-    from core.progression import hit_point_bonus_sources_from_grants
+    from core.hp_bonus import hit_point_bonus_sources_from_grants
 
     return hit_point_bonus_sources_from_grants(
         collect_race_grants(race_id, subrace_id)
